pyproject.toml
[project]
name = "pilot-deep-agents"
version = "0.1.0"
description = "Deep Agents example (LangChain-based, OpenAI-compatible API)"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "deepagents>=0.4.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.9",
    "langchain-core>=1.2.11",
    "langgraph>=1.0.8",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "ty>=0.0.16",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]

[tool.ty]
environment = { python-version = "3.12" }

[tool.ty.rules]
# Third-party API args (e.g. LangChain) not fully recognized
unknown-argument = "ignore"
